+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/FDBRecordStore.java
@@ -3529,7 +3529,18 @@ public class FDBRecordStore extends FDBStoreBase implements FDBRecordStoreBase<M
      * Reason that an index is being rebuilt now.
      */
     public enum RebuildIndexReason {
-        NEW_STORE, FEW_RECORDS, COUNTS_UNKNOWN, REBUILD_ALL, EXPLICIT, TEST
+        NEW_STORE(FDBStoreTimer.Events.REBUILD_INDEX_NEW_STORE),
+        FEW_RECORDS(FDBStoreTimer.Events.REBUILD_INDEX_FEW_RECORDS),
+        COUNTS_UNKNOWN(FDBStoreTimer.Events.REBUILD_INDEX_COUNTS_UNKNOWN),
+        REBUILD_ALL(FDBStoreTimer.Events.REBUILD_INDEX_REBUILD_ALL),
+        EXPLICIT(FDBStoreTimer.Events.REBUILD_INDEX_EXPLICIT),
+        TEST(FDBStoreTimer.Events.REBUILD_INDEX_TEST);
+
+        public final FDBStoreTimer.Events event;
+
+        RebuildIndexReason(FDBStoreTimer.Events event) {
+            this.event = event;
+        }
     }
 
     private boolean areAllRecordTypesSince(@Nullable Collection<RecordType> recordTypes, @Nullable Integer oldMetaDataVersion) {
@@ -3647,7 +3658,7 @@ public class FDBRecordStore extends FDBStoreBase implements FDBRecordStoreBase<M
                     return null;
                 });
 
-        return context.instrument(FDBStoreTimer.Events.REBUILD_INDEX, future, startTime);
+        return context.instrument(reason.event, future, startTime);
     }
 
     @SuppressWarnings("PMD.GuardLogStatement") // Already is, but around several call.